requirements.txt
ansible
six
junos-eznc
jxmlease
python-dotenv
urllib3
requests
